+++ framework/memory.h
@@ -8,18 +8,18 @@ class Memory {
 public:
     static void* malloctrack(size_t size);
     static void freetrack(void* ptr);
-    static int getCurrent();
-    static int getMax();
-    static int getTotal();
+    static long getCurrent();
+    static long getMax();
+    static long getTotal();
     static void printCurrent();
     static void printMax();
     static void printTotal();
     static void printAll();
 
 private:
-    static int current;
-    static int max;
-    static int total;
+    static long current;
+    static long max;
+    static long total;
 };
 
 #endif
